requirements.txt
Flask
google-cloud-storage
google-cloud-bigquery
google-cloud-aiplatform
PyGithub
gunicorn